Making sure that the build version is stable to be deploy it
