+++ src/lib/authProviders.ts
@@ -1,0 +1,19 @@
+// src/lib/auth.ts
+import { NextAuthOptions } from "next-auth";
+import GithubProvider from "next-auth/providers/github";
+import GoogleProvider from "next-auth/providers/google";
+
+export const authOptions: NextAuthOptions = {
+  providers: [
+    GoogleProvider({
+      clientId: process.env.GOOGLE_CLIENT_ID || "",
+      clientSecret: process.env.GOOGLE_CLIENT_SECRET || "",
+    }),
+    GithubProvider({
+      clientId: process.env.GITHUB_CLIENT_ID || "",
+      clientSecret: process.env.GITHUB_CLIENT_SECRET || "",
+      authorization: { params: { scope: "read:user user:email" } },
+    }),
+  ],
+  secret: process.env.AUTH_SECRET,
+};
